Moved opportunity routes to opportunity routes.py in order to organize the code and have better maintenance. Added location select menu to create page for opportunities
