+++ labconnect/main/routes.py
@@ -25,121 +25,9 @@ from labconnect.models import (
     RecommendsCourses,
     RecommendsClassYears,
 )
+from labconnect.helpers import LocationEnum, SemesterEnum
 
 from . import main_blueprint
-
-
-def packageOpportunity(opportunityInfo, professorInfo):
-    data = opportunityInfo.to_dict()
-    data["professor"] = professorInfo.name
-    data["department"] = professorInfo.department_id
-
-    return data
-
-
-def packageIndividualOpportunity(opportunityInfo):
-    data = {}
-    data["id"] = opportunityInfo.id
-    data["name"] = opportunityInfo.name
-    data["description"] = opportunityInfo.description
-    data["recommended_experience"] = opportunityInfo.recommended_experience
-    data["author"] = ""
-    data["department"] = ""
-
-    credits = ""
-    if opportunityInfo.one_credit:
-        credits += "1, "
-    if opportunityInfo.two_credits:
-        credits += "2, "
-    if opportunityInfo.three_credits:
-        credits += "3, "
-    if opportunityInfo.four_credits:
-        credits += "4"
-
-    if credits != "":
-        credits += " credits"
-
-    data["aboutSection"] = [
-        {
-            "title": "Pay",
-            "description": f"${opportunityInfo.pay} per hour",
-        },
-        {
-            "title": "Semester",
-            "description": f"{opportunityInfo.semester} {opportunityInfo.year}",
-        },
-        {
-            "title": "Application Due",
-            "description": opportunityInfo.application_due,
-        },
-    ]
-
-    if credits != "":
-        data["aboutSection"].append(
-            {
-                "title": "Credits",
-                "description": credits,
-            }
-        )
-
-    # get professor and department by getting Leads and LabManager
-
-    query = db.session.execute(
-        db.select(Leads, LabManager)
-        .filter(Leads.opportunity_id == opportunityInfo.id)
-        .join(LabManager, Leads.lab_manager_rcs_id == LabManager.rcs_id)
-    )
-
-    queryInfo = query.all()
-    print(queryInfo)
-
-    if len(queryInfo) == 0:
-        return data
-
-    data["department"] = queryInfo[0][1].department_id
-
-    for i in range(len(queryInfo)):
-        data["author"] += queryInfo[i][1].name
-        if i != len(queryInfo) - 1:
-            data["author"] += ", "
-
-    if len(queryInfo) > 1:
-        data["authorProfile"] = "https://t4.ftcdn.net/jpg/03/78/40/51/360_F_378405187_PyVLw51NVo3KltNlhUOpKfULdkUOUn7j.jpg"
-    elif len(queryInfo) == 1:
-        data["authorProfile"] = "https://cdn-icons-png.flaticon.com/512/1077/1077114.png"
-
-    return data
-
-
-def packageOpportunityCard(opportunity):
-
-    # get professor and department by getting Leads and LabManager
-    query = db.session.execute(
-        db.select(Leads, LabManager)
-        .filter(Leads.opportunity_id == opportunity.id)
-        .join(LabManager, Leads.lab_manager_rcs_id == LabManager.rcs_id)
-    )
-
-    data = query.all()
-
-    professorInfo = ""
-
-    for i in range(len(data)):
-        professorInfo += data[i][1].name
-        if i != len(data) - 1:
-            professorInfo += ", "
-
-    card = {
-        "id": opportunity.id,
-        "title": opportunity.name,
-        "professor": professorInfo,
-        "season": opportunity.semester,
-        "location": "TBA",
-        "year": opportunity.year,
-    }
-
-    return card
-
 
 @main_blueprint.route("/")
 def index():
@@ -350,26 +238,7 @@ def getProfessorProfile(rcs_id: str):
     return dictionary
 
 
-@main_blueprint.route("/getOpportunity/<int:opp_id>", methods=["GET"])
-def getOpportunity(opp_id: int):
-    # query database for opportunity
-    query = db.session.execute(
-        db.select(Opportunities)
-        .filter(Opportunities.id == opp_id)
-    )
 
-    data = query.all()
-
-    # check if opportunity exists
-    if not data or len(data) == 0:
-        abort(404)
-
-    data = data[0]
-
-    oppData = packageIndividualOpportunity(data[0])
-
-    # return data in the below format if opportunity is found
-    return {"data": oppData}
 
 
 @main_blueprint.get("/lab_manager/opportunities")
@@ -398,164 +267,14 @@ def getLabManagerOpportunityCards() -> dict[Any, list[Any]]:
     return result
 
 
-@main_blueprint.route("/getProfessorOpportunityCards/<string:rcs_id>", methods=["GET"])
-def getProfessorOpportunityCards(rcs_id: str):
-    if request.method == "GET":
-        # query database for opportunity
-
-        query = db.session.execute(
-            db.select(Opportunities, Leads)
-            .filter(Leads.lab_manager_rcs_id == rcs_id)
-            .join(Opportunities, Leads.opportunity_id == Opportunities.id)
-        )
-
-        data = query.all()
-
-        cards = {"data": []}
-
-        for row in data:
-            opportunity = row[0]
-
-            if not opportunity.active:
-                continue
-
-            oppData = {
-                "id": opportunity.id,
-                "title": opportunity.name,
-                "body": "Due " + str(opportunity.application_due),
-                "attributes": [],
-            }
-
-            if opportunity.pay != None and opportunity.pay > 0:
-                oppData["attributes"].append("Paid")
-
-            if (
-                opportunity.one_credit
-                or opportunity.two_credits
-                or opportunity.three_credits
-                or opportunity.four_credits
-            ):
-                oppData["attributes"].append("Credit Available")
-
-            cards["data"].append(oppData)
-
-        # return data in the below format if opportunity is found
-        return cards
-
-    abort(500)
-
-
-@main_blueprint.route("/getProfileOpportunities/<string:rcs_id>", methods=["GET"])
-def getProfileOpportunities(rcs_id: str):
-    if request.method == "GET":
-        # query database for opportunity
-
-        query = db.session.execute(
-            db.select(Opportunities, Leads)
-            .filter(Leads.lab_manager_rcs_id == rcs_id)
-            .join(Opportunities, Leads.opportunity_id == Opportunities.id)
-        )
-
-        data = query.all()
-
-        cards = {"data": []}
-
-        for row in data:
-            opportunity = row[0]
-
-            oppData = {
-                "id": opportunity.id,
-                "title": opportunity.name,
-                "body": "Due " + str(opportunity.application_due),
-                "attributes": [],
-                "activeStatus": opportunity.active,
-            }
-
-            if opportunity.pay != None and opportunity.pay > 0:
-                oppData["attributes"].append("Paid")
-            if (
-                opportunity.one_credit
-                or opportunity.two_credits
-                or opportunity.three_credits
-                or opportunity.four_credits
-            ):
-                oppData["attributes"].append("Credits")
-
-            cards["data"].append(oppData)
-
-        # return data in the below format if opportunity is found
-        return cards
-
-    abort(500)
-
-
-@main_blueprint.route("/getOpportunityByProfessor/<string:rcs_id>", methods=["GET"])
-def getOpportunityByProfessor(rcs_id: str):
-    if request.method == "GET":
-        # query database for opportunity
-        query = db.session.execute(
-            db.select(Opportunities, Leads)
-            .filter(Leads.lab_manager_rcs_id == rcs_id)
-            .join(Opportunities, Leads.opportunity_id == Opportunities.id)
-        )
-
-        data = query.all()
-        print(data)
-
-        # return data in the below format if opportunity is found
-        return {"data": [opportunity[0].to_dict() for opportunity in data]}
-
-    abort(500)
-
-
-@main_blueprint.route("/getOpportunities", methods=["GET"])
-def getOpportunities():
-    if request.method == "GET":
-        # query database for opportunity
-        query = db.session.execute(
-            db.select(Opportunities, Leads, LabManager)
-            .join(Leads, Leads.opportunity_id == Opportunities.id)
-            .join(LabManager, Leads.lab_manager_rcs_id == LabManager.rcs_id)
-        )
-        data = query.all()
-        print(data[0])
-
-        # return data in the below format if opportunity is found
-        return {
-            "data": [
-                packageOpportunity(opportunity[0], opportunity[2])
-                for opportunity in data
-            ]
-        }
-
-    abort(500)
-
-
-# Jobs page
-
-
-@main_blueprint.route("/getOpportunityCards", methods=["GET"])
-def getOpportunityCards():
-    if request.method == "GET":
-        # query database for opportunity
-        query = db.session.execute(
-            db.select(Opportunities)
-            .filter(Opportunities.active == True)
-        )
-
-        data = query.fetchall()
-
-        # return data in the below format if opportunity is found
-        cards = {
-            "data": [packageOpportunityCard(opportunity[0]) for opportunity in data]
-        }
-
-        return cards
 
 
 
 
-    abort(500)
+
+
+
+
 
 
 # _______________________________________________________________________________________________#
@@ -584,119 +303,7 @@ def getProfessorCookies(id: str):
     return dictionary
 
 
-@main_blueprint.route("/getOpportunityMeta/<int:id>", methods=["GET"])
-def getOpportunityMeta(id: int):
-    if request.method == "GET":
-        query = db.session.execute(
-            db.select(
-                Opportunities, RecommendsMajors, RecommendsCourses, RecommendsClassYears
-            )
-            .filter(Opportunities.id == id)
-            .join(RecommendsMajors, RecommendsMajors.opportunity_id == Opportunities.id)
-            .join(
-                RecommendsCourses, RecommendsCourses.opportunity_id == Opportunities.id
-            )
-            .join(
-                RecommendsClassYears,
-                RecommendsClassYears.opportunity_id == Opportunities.id,
-            )
-        )
-        data = query.all()
-        print(data)
 
-        if not data or len(data) == 0:
-            abort(404)
-
-        dictionary = data[0][0].to_dict()
-        dictionary["semester"] = dictionary["semester"].upper()
-        dictionary["courses"] = set()
-        dictionary["majors"] = set()
-        dictionary["years"] = set()
-
-        for row in data:
-            dictionary["courses"].add(row[2].course_code)
-            dictionary["majors"].add(row[1].major_code)
-            dictionary["years"].add(row[3].class_year)
-
-        dictionary["courses"] = list(dictionary["courses"])
-        dictionary["majors"] = list(dictionary["majors"])
-        dictionary["years"] = list(dictionary["years"])
-
-        for i in range(len(dictionary["years"])):
-            dictionary["years"][i] = str(dictionary["years"][i])
-
-        dictionary["credits"] = []
-        if dictionary["one_credit"]:
-            dictionary["credits"].append("1")
-
-        if dictionary["two_credits"]:
-            dictionary["credits"].append("2")
-
-        if dictionary["three_credits"]:
-            dictionary["credits"].append("3")
-
-        if dictionary["four_credits"]:
-            dictionary["credits"].append("4")
-
-        dictionary.pop("one_credit")
-        dictionary.pop("two_credits")
-        dictionary.pop("three_credits")
-        dictionary.pop("four_credits")
-
-        return {"data": dictionary}
-
-    abort(500)
-
-
-@main_blueprint.route("/deleteOpportunity", methods=["DELETE", "POST"])
-def deleteOpportunity():
-    if request.method in ["DELETE", "POST"]:
-        data = request.json
-        id = data["id"]
-
-        query = db.session.execute(
-            db.select(
-                Opportunities,
-                RecommendsMajors,
-                RecommendsCourses,
-                RecommendsClassYears,
-                Leads,
-            )
-            .filter(Opportunities.id == id)
-            .join(RecommendsMajors, RecommendsMajors.opportunity_id == Opportunities.id)
-            .join(
-                RecommendsCourses, RecommendsCourses.opportunity_id == Opportunities.id
-            )
-            .join(
-                RecommendsClassYears,
-                RecommendsClassYears.opportunity_id == Opportunities.id,
-            )
-            .join(Leads, Leads.opportunity_id == Opportunities.id)
-        )
-
-        data = query.all()
-        print(data)
-
-        if not data or len(data) == 0:
-            abort(404)
-
-        opportunity = data[0][0]
-
-        for row in data:
-            db.session.delete(row[1])
-            db.session.delete(row[2])
-            db.session.delete(row[3])
-            db.session.delete(row[4])
-
-        leads = data[0][4]
-
-        db.session.delete(opportunity)
-
-        db.session.commit()
-
-        return "Success"
-
-    abort(500)
 
 
 @main_blueprint.route("/changeActiveStatus", methods=["DELETE", "POST"])
@@ -730,212 +337,6 @@ def changeActiveStatus():
         # if match is found, change the opportunities active status to true or false based on setStatus
 
         return {"activeStatus": data.active}
-
-    abort(500)
-
-
-@main_blueprint.route("/editOpportunity", methods=["DELETE", "POST"])
-def editOpportunity():
-    if True:
-        data = request.json
-        id = data["id"]
-        # authToken = data["authToken"]
-        # authorID = data["authorID"]
-        newPostData = data
-
-        # query database to see if the credentials above match
-        query = db.session.execute(
-            db.select(
-                Opportunities, RecommendsMajors, RecommendsCourses, RecommendsClassYears
-            )
-            .filter(Opportunities.id == id)
-            .join(RecommendsMajors, RecommendsMajors.opportunity_id == Opportunities.id)
-            .join(
-                RecommendsCourses, RecommendsCourses.opportunity_id == Opportunities.id
-            )
-            .join(
-                RecommendsClassYears,
-                RecommendsClassYears.opportunity_id == Opportunities.id,
-            )
-        )
-
-        data = query.all()
-
-        if not data or len(data) == 0:
-            abort(404)
-
-        opportunity = data[0][0]
-
-        one = False
-        two = False
-        three = False
-        four = False
-
-        if "1" in newPostData["credits"]:
-            one = True
-        if "2" in newPostData["credits"]:
-            two = True
-        if "3" in newPostData["credits"]:
-            three = True
-        if "4" in newPostData["credits"]:
-            four = True
-
-        # if match is found, edit the opportunity with the new data provided
-        opportunity.name = newPostData["name"]
-        opportunity.description = newPostData["description"]
-        opportunity.recommended_experience = newPostData["recommended_experience"]
-        opportunity.pay = newPostData["pay"]
-        opportunity.one_credit = one
-        opportunity.two_credits = two
-        opportunity.three_credits = three
-        opportunity.four_credits = four
-        opportunity.semester = newPostData["semester"]
-        opportunity.year = newPostData["year"]
-        opportunity.application_due = datetime.datetime.strptime(
-            newPostData["application_due"], "%Y-%m-%d"
-        )
-        opportunity.active = newPostData["active"]
-        db.session.add(opportunity)
-        db.session.commit()
-
-        # delete all the old data in the recommends tables
-
-        for row in data:
-            db.session.delete(row[1])
-            db.session.delete(row[2])
-            db.session.delete(row[3])
-
-        # create new data for allow the tables
-
-        for course in newPostData["courses"]:
-            newCourse = RecommendsCourses(
-                opportunity_id=opportunity.id, course_code=course
-            )
-            db.session.add(newCourse)
-            db.session.commit()
-
-        for major in newPostData["majors"]:
-            newMajor = RecommendsMajors(opportunity_id=opportunity.id, major_code=major)
-            db.session.add(newMajor)
-            db.session.commit()
-
-        for year in newPostData["years"]:
-            newYear = RecommendsClassYears(
-                opportunity_id=opportunity.id, class_year=year
-            )
-            db.session.add(newYear)
-            db.session.commit()
-
-        return "Successful"
-
-    abort(500)
-
-
-# create fake opportunity
-"""
-{
-    "authorID": "led",
-    "newPostData": {
-        "name": "Abid's AC System",
-        "description": "Time to cool your room with a new AC system!",
-        "recommended_experience": "Computer Science, ITWS, or any other major with a focus on technology.",
-        "pay": 25.0,
-        "credits": "4",
-        "semester": "Fall",
-        "year": 2024,
-        "application_due": "2024-03-30",
-        "active": True,
-        "courses": ["CSCI4430"],
-        "majors": ["PHYS"],
-        "years": [2023, 2024]
-    }
-}
-
-"""
-
-
-@main_blueprint.route("/createOpportunity", methods=["POST"])
-def createOpportunity():
-    if request.method == "POST":
-        data = request.json
-        authorID = data["authorID"]
-        newPostData = data
-
-        # query database to see if the credentials above match
-        query = db.session.execute(
-            db.select(LabManager).filter(LabManager.rcs_id == authorID)
-        )
-
-        data = query.all()[0][0]
-
-        # TODO: how do we get the opportunity id?
-        # if match is found, create a new opportunity with the new data provided
-
-        one = False
-        two = False
-        three = False
-        four = False
-
-        if "1" in newPostData["credits"]:
-            one = True
-        if "2" in newPostData["credits"]:
-            two = True
-        if "3" in newPostData["credits"]:
-            three = True
-        if "4" in newPostData["credits"]:
-            four = True
-
-        newOpportunity = Opportunities(
-            name=newPostData["name"],
-            description=newPostData["description"],
-            recommended_experience=newPostData["recommended_experience"],
-            pay=newPostData["pay"],
-            one_credit=one,
-            two_credits=two,
-            three_credits=three,
-            four_credits=four,
-            semester=newPostData["semester"],
-            year=newPostData["year"],
-            application_due=datetime.datetime.strptime(
-                newPostData["application_due"], "%Y-%m-%d"
-            ),
-            active=newPostData["active"],
-        )
-        print("before comitting")
-        db.session.add(newOpportunity)
-        db.session.commit()
-
-        print("got here atleast")
-
-        newLead = Leads(lab_manager_rcs_id=authorID, opportunity_id=newOpportunity.id)
-
-        db.session.add(newLead)
-        db.session.commit()
-
-        for course in newPostData["courses"]:
-            newCourse = RecommendsCourses(
-                opportunity_id=newOpportunity.id, course_code=course
-            )
-            db.session.add(newCourse)
-            db.session.commit()
-
-        for major in newPostData["majors"]:
-            newMajor = RecommendsMajors(
-                opportunity_id=newOpportunity.id, major_code=major
-            )
-            db.session.add(newMajor)
-            db.session.commit()
-
-        for year in newPostData["years"]:
-            newYear = RecommendsClassYears(
-                opportunity_id=newOpportunity.id, class_year=year
-            )
-            db.session.add(newYear)
-            db.session.commit()
-
-        # db.session.add(newOpportunity)
-
-        return {"data": "Opportunity Created"}
 
     abort(500)
